refactor: 배경 그라데이션 위치 수정 및 부모 영역 패딩 추가
