+++ client/src/widgets/albums/index.ts
@@ -3,4 +3,3 @@ export { Comment } from './ui/Comment.tsx';
 export { CommentList } from './ui/CommentList.tsx';
 export { Playlist } from './ui/Playlist';
 export { TrackItem } from './ui/TrackItem';
-export { AlbumDetailBackground } from './ui/AlbumDetailBackground.tsx';
